Created multiple breeds for each age group. Added school (donde = 3, lugar = 3).
Before going any further, I tried running the simulation in NetLogo, and I'm getting the following error: "AND expected input to be a TRUE/FALSE but got the number 0 instead." I can't seem to find where in the code it might be defaulting to 0, when it should be FALSE (assuming that's the problem). Any ideas?